Modified Home route to /health and added a file for general controllers.
